+++ test/spec/Feature/Query/QuerySpec.hs
@@ -7,6 +7,8 @@ import Test.Hspec          hiding (pendingWith)
 import Test.Hspec.Wai
 import Test.Hspec.Wai.JSON
 
+import PostgREST.Config (AppConfig (..))
+
 import Protolude  hiding (get)
 import SpecHelper
 
@@ -1069,15 +1071,15 @@ spec withConfig = withConfig baseCfg $ do
             { "id":4,"children":[]}
           ]|] { matchHeaders = [matchContentTypeJson] }
 
-      it "works when embedding the same table more than once" $
-        get "/places?select=name,visits(id,start_time,visit_type),work_visits:visits(id,start_time,visit_type)&id=eq.1&visits.visit_type=neq.work&visits.start_time=gt.20250101+00:00&work_visits.visit_type=eq.work&work_visits.start_time=gt.20250101+00:00" `shouldRespondWith`
-          [json|[
-            {
-              "name":"Lake",
-              "visits":[{"id": 1, "start_time": "2025-01-01T10:00:00", "visit_type": "vacation"}, {"id": 2, "start_time": "2025-01-01T15:00:00", "visit_type": "vacation"}],
-              "work_visits":[{"id": 3, "start_time": "2025-01-01T20:00:00", "visit_type": "work"}]
-            }
-          ]|] { matchHeaders = [matchContentTypeJson] }
+    it "works when embedding the same table more than once" $
+      get "/places?select=name,visits(id,start_time,visit_type),work_visits:visits(id,start_time,visit_type)&id=eq.1&visits.visit_type=neq.work&visits.start_time=gt.20250101+00:00&work_visits.visit_type=eq.work&work_visits.start_time=gt.20250101+00:00" `shouldRespondWith`
+        [json|[
+          {
+            "name":"Lake",
+            "visits":[{"id": 1, "start_time": "2025-01-01T10:00:00", "visit_type": "vacation"}, {"id": 2, "start_time": "2025-01-01T15:00:00", "visit_type": "vacation"}],
+            "work_visits":[{"id": 3, "start_time": "2025-01-01T20:00:00", "visit_type": "work"}]
+          }
+        ]|] { matchHeaders = [matchContentTypeJson] }
 
   describe "ordering response" $ do
     it "by a column asc" $
@@ -1164,6 +1166,19 @@ spec withConfig = withConfig baseCfg $ do
       get "/projects?id=eq.1&select=id, name, the_tasks:tasks(id, name)&the_tasks.order=name.asc" `shouldRespondWith`
         [json|[{"id":1,"name":"Windows 7","the_tasks":[{"id":2,"name":"Code w7"},{"id":1,"name":"Design w7"}]}]|]
         { matchHeaders = [matchContentTypeJson] }
+
+    it "does not work when ordering embedded entities with alias using the target name" $
+      get "/projects?id=eq.1&select=id,name,the_tasks:tasks(id,name)&tasks.order=name.asc" `shouldRespondWith`
+        [json|
+          {
+            "code":"PGRST108",
+            "details":"Target names are not allowed in filters if they have an alias",
+            "hint":"Change 'tasks' to 'the_tasks' in filters, orders or limits.",
+            "message":"'tasks' is not an embedded resource in this request"
+          }
+        |]
+        { matchStatus = 400,
+          matchHeaders = [matchContentTypeJson] }
 
     it "ordering embeded entities, two levels" $
       get "/projects?id=eq.1&select=id, name, tasks(id, name, users(id, name))&tasks.order=name.asc&tasks.users.order=name.desc" `shouldRespondWith`
@@ -1685,3 +1700,11 @@ spec withConfig = withConfig baseCfg $ do
       [json| {"code":"PGRST125","details":null,"hint":null,"message":"Invalid path specified in request URL"} |]
       { matchStatus = 404
       , matchHeaders = ["Content-Length" <:> "96"]}
+
+specLegacyTargetNames :: SpecWithConfig
+specLegacyTargetNames withConfig = withConfig (baseCfg { configUrlUseLegacyTargetNames = True }) $
+  context "enable legacy target names" $
+    it "filters, orders and limits embeded entities with alias using the target name" $
+      get "/projects?id=eq.1&select=id, name, the_tasks:tasks(id, name)&tasks.name=like.Code*&tasks.order=name.asc&tasks.limit=1" `shouldRespondWith`
+        [json|[{"id":1,"name":"Windows 7","the_tasks":[{"id":2,"name":"Code w7"}]}]|]
+        { matchHeaders = ["Warning" <:> "299 PostgRESTv15(pre-release) \"Embedded resource was referenced by relation name even though it has an alias. This is deprecated and will stop working in a future release. Update `tasks` to `the_tasks` in query string filters, orders or limits.\""] }