+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to this project will be documented in this file. From versio
 - Log schema cache queries timings on `log-level=debug` by @steve-chavez in #4805
 - Add GHC runtime metrics to the metrics endpoint by @mkleczek in #4862
 - Support running the admin server on a unix socket by @wolfgangwalther in #5003
+- Add config `url-use-legacy-target-names` to allow using target names in filters, orders or limits when it has an alias by @laurenceisla in #4075
 
 ### Fixed
 
@@ -32,8 +33,9 @@ All notable changes to this project will be documented in this file. From versio
   + Now fails at startup. Prior to this, it failed with `PGRST205` on requests related to these schemas.
 - Build a static executable for aarch64-linux by @wolfgangwalther in #4193
 - Build the minimal docker image for aarch64-linux by @wolfgangwalther in #4193
-- The name of an embedded table can no longer be used in filters if it has an alias by @laurenceisla in #4075
+- The name of an embedded table can no longer be used in filters, orders or limits if it has an alias by @steve-chavez, @laurenceisla in #4075
   + e.g. `?select=alias:table(*)&table.id=eq.1` is not possible anymore, use `?select=alias:table(*)&alias.id=eq.1` instead.
+  + You can bypass this breaking change by setting `url-use-legacy-target-names = true` (you will see a warning in the logs when this happens)
 - Config `jwt-role-claim-key` now uses RFC 9535 syntax for JSON Path by @taimoorzaeem in #4984
 
 #### Changed Syntax for JWT Role Extraction
@@ -47,6 +49,10 @@ The `jwt-role-claim-key` config should be updated according to the following:
 - String comparison operators (`^==`, `==^` and `*==`) are replaced with regular expression search.
   + Example: `.roles[?(@ ^== "postgrest_test_")]` -> `$.roles[?search(@, "^postgrest_test_")]`
 - Detailed reference for syntax: [RFC 9535](https://www.rfc-editor.org/rfc/rfc9535.html#name-jsonpath-syntax-and-semanti).
+
+### Deprecated
+
+- Deprecate bypassing breaking change using `url-use-legacy-target-names = true` by @steve-chavez, @laurenceisla in #4075
 
 ## [14.14] - 2026-06-29